Cambios
- Se cambia el header de la pagina
- Se colocan los beneficios del usuario
- Se coloca que cuando se una el usuario a un grupo de interes ya no lea aparezca el boton de unirse
- Se colocan los beneficios y puntos en usuario gestor y usuario administrador
- Se hace la tabla de usuarios comunidad
- Se hace el dao y dto de la usuarios comunidad
- Se hace el controlador y service de usuarios comunidad

+++ system/php/modules/answerQuestion/ServiceAnswerQuestion.php
@@ -8,21 +8,23 @@ class ServiceAnswerQuestion extends System
     public static function newAnswerQuestion($id_encuesta, $id_pregunta, $respuesta, $veracidad)
     {
         try {
-            $id_encuesta    = parent::limpiarString($id_encuesta);
-            $id_pregunta    = parent::limpiarString($id_pregunta);
-            $respuesta      = parent::limpiarString($respuesta);
-            $veracidad      = parent::limpiarString($veracidad);
-            $fecha_registro = date('Y-m-d H:i:s');
+            if (basename($_SERVER['PHP_SELF']) == 'surveyQuestion.php') {
+                $id_encuesta    = parent::limpiarString($id_encuesta);
+                $id_pregunta    = parent::limpiarString($id_pregunta);
+                $respuesta      = parent::limpiarString($respuesta);
+                $veracidad      = parent::limpiarString($veracidad);
+                $fecha_registro = date('Y-m-d H:i:s');
 
-            $validateCorrect = RespuestaPregunta::validateOnlyAnswerCorrect($id_pregunta);
-            if (!$validateCorrect || $veracidad == 0) {
-                $result = RespuestaPregunta::newAnswerQuestion($id_encuesta, $id_pregunta, $respuesta,  $veracidad, $fecha_registro);
+                $validateCorrect = RespuestaPregunta::validateOnlyAnswerCorrect($id_pregunta);
+                if (!$validateCorrect || $veracidad == 0) {
+                    $result = RespuestaPregunta::newAnswerQuestion($id_encuesta, $id_pregunta, $respuesta,  $veracidad, $fecha_registro);
 
-                if ($result) {
-                    return Elements::crearMensajeAlerta(Constants::$REGISTER_NEW, "success");
+                    if ($result) {
+                        return Elements::crearMensajeAlerta(Constants::$REGISTER_NEW, "success");
+                    }
+                } else {
+                    return Elements::crearMensajeAlerta(Constants::$ANSWER_REPEAT, "error");
                 }
-            } else {
-                return Elements::crearMensajeAlerta(Constants::$ANSWER_REPEAT, "error");
             }
         } catch (\Exception $e) {
             throw new Exception($e->getMessage());
@@ -31,14 +33,16 @@ class ServiceAnswerQuestion extends System
     public static function setAnswerQuestion($id_respuesta, $respuesta, $veracidad)
     {
         try {
-            $id_respuesta  = parent::limpiarString($id_respuesta);
-            $respuesta     = parent::limpiarString($respuesta);
-            $veracidad     = parent::limpiarString($veracidad);
+            if (basename($_SERVER['PHP_SELF']) == 'answerQuestion.php') {
+                $id_respuesta  = parent::limpiarString($id_respuesta);
+                $respuesta     = parent::limpiarString($respuesta);
+                $veracidad     = parent::limpiarString($veracidad);
 
-            $result = RespuestaPregunta::setAnswerQuestion($id_respuesta, $respuesta, $veracidad);
+                $result = RespuestaPregunta::setAnswerQuestion($id_respuesta, $respuesta, $veracidad);
 
-            if ($result) {
-                return Elements::crearMensajeAlerta(Constants::$REGISTER_UPDATE, "success");
+                if ($result) {
+                    return Elements::crearMensajeAlerta(Constants::$REGISTER_UPDATE, "success");
+                }
             }
         } catch (\Exception $e) {
             throw new Exception($e->getMessage());
@@ -57,14 +61,16 @@ class ServiceAnswerQuestion extends System
     public static function deleteAnswerQuestion($id_respuesta)
     {
         try {
-            $id_respuesta = parent::limpiarString($id_respuesta);
-            $respuestaDTO = RespuestaPregunta::getAnswerQuestion($id_respuesta);
-            $result = RespuestaPregunta::deleteAnswerQuestion($id_respuesta);
-            if ($result) {
-                $_SESSION['alert'] = 1;
-                header('Location:surveyQuestion?surveyQuestion=' . $respuestaDTO->getPreguntaDTO()->getId_pregunta() . '&survey=' . $respuestaDTO->getId_respuesta() . '&delete');
-            } else {
-                return Elements::crearMensajeAlerta(Constants::$REGISTER_DELETE_NOT, "error");
+            if (basename($_SERVER['PHP_SELF']) == 'answerQuestion.php') {
+                $id_respuesta = parent::limpiarString($id_respuesta);
+                $respuestaDTO = RespuestaPregunta::getAnswerQuestion($id_respuesta);
+                $result = RespuestaPregunta::deleteAnswerQuestion($id_respuesta);
+                if ($result) {
+                    $_SESSION['alert'] = 1;
+                    header('Location:surveyQuestion?surveyQuestion=' . $respuestaDTO->getPreguntaDTO()->getId_pregunta() . '&survey=' . $respuestaDTO->getId_respuesta() . '&delete');
+                } else {
+                    return Elements::crearMensajeAlerta(Constants::$REGISTER_DELETE_NOT, "error");
+                }
             }
         } catch (\Exception $e) {
             throw new Exception($e->getMessage());
@@ -97,23 +103,25 @@ class ServiceAnswerQuestion extends System
     public static function getCardsQuetionAndAnwer($id_encuesta)
     {
         try {
-            $id_encuesta = parent::limpiarString($id_encuesta);
-            $preguntaDTO = PreguntaEncuesta::listSurveyQuestionBySurveyByEstate($id_encuesta);
-            $cardHtml = '';
-            $con = 0;
-            foreach ($preguntaDTO as $value) {
-                $html = '';
-                $respuestaDTO = RespuestaPregunta::listAnswerQuestionByQuestion($value->getId_pregunta());
-                foreach ($respuestaDTO as $valor) {
-                    $html .= '<div class="form-check row">
+            if (basename($_SERVER['PHP_SELF']) == 'startSurvey.php') {
+                $id_encuesta = parent::limpiarString($id_encuesta);
+                $preguntaDTO = PreguntaEncuesta::listSurveyQuestionBySurveyByEstate($id_encuesta);
+                $cardHtml = '';
+                $con = 0;
+                foreach ($preguntaDTO as $value) {
+                    $html = '';
+                    $respuestaDTO = RespuestaPregunta::listAnswerQuestionByQuestion($value->getId_pregunta());
+                    foreach ($respuestaDTO as $valor) {
+                        $html .= '<div class="form-check row">
                                 <input type="radio" name="' . $value->getId_pregunta() . '" value="' . $valor->getId_respuesta() . '" class="form-check-input col-1">
                                 <label for="' . $value->getId_pregunta() . '" class="form-check-label col-11" style="color:black; font-size:16px;">' . $valor->getRespuesta() . '</label>
                             </div>';
+                    }
+                    $con++;
+                    $cardHtml .= Elements::getCardQuestion($con, $value->getPregunta(), $html);
                 }
-                $con++;
-                $cardHtml .= Elements::getCardQuestion($con, $value->getPregunta(), $html);
+                return $cardHtml;
             }
-            return $cardHtml;
         } catch (\Exception $e) {
             throw new Exception($e->getMessage());
         }
